move bonds.txt parsing logic to BondsParser
