+++ casper/src/main/scala/coop/rchain/casper/genesis/Genesis.scala
@@ -1,27 +1,19 @@
 package coop.rchain.casper.genesis
 
-import java.io.{FileNotFoundException, PrintWriter}
-import java.nio.file.{Path, Paths}
+import java.nio.file.Path
 
 import cats.effect.{Concurrent, Sync}
 import cats.implicits._
-import cats.{Foldable, Monad}
 import com.google.protobuf.ByteString
 import coop.rchain.blockstorage.util.io.IOError.RaiseIOError
-import coop.rchain.blockstorage.util.io._
 import coop.rchain.casper.genesis.contracts._
 import coop.rchain.casper.protocol._
 import coop.rchain.casper.util.ProtoUtil.{blockHeader, unsignedBlockProto}
 import coop.rchain.casper.util.Sorting.byteArrayOrdering
-import coop.rchain.casper.util.VaultParser
 import coop.rchain.casper.util.rholang.RuntimeManager.StateHash
 import coop.rchain.casper.util.rholang.{InternalProcessedDeploy, RuntimeManager}
-import coop.rchain.crypto.PublicKey
-import coop.rchain.crypto.codec.Base16
-import coop.rchain.crypto.signatures.Secp256k1
+import coop.rchain.casper.util.{BondsParser, VaultParser}
 import coop.rchain.shared.{Log, LogSource, Time}
-
-import scala.util.{Failure, Success, Try}
 
 final case class Genesis(
     shardId: String,
@@ -32,8 +24,6 @@ final case class Genesis(
 )
 
 object Genesis {
-
-  implicit private val logSource: LogSource = LogSource(this.getClass)
 
   def defaultBlessedTerms(
       timestamp: Long,
@@ -69,7 +59,7 @@ object Genesis {
     for {
       timestamp <- deployTimestamp.fold(Time[F].currentMillis)(_.pure[F])
       vaults    <- VaultParser.parse[F](maybeVaultsPath, genesisPath.resolve("wallets.txt"))
-      bonds <- getBonds[F](
+      bonds <- BondsParser.parse[F](
                 maybeBondsPath,
                 genesisPath.resolve("bonds.txt"),
                 numValidators,
@@ -122,7 +112,6 @@ object Genesis {
     import genesis._
 
     val state = RChainState(
-      blockNumber = 0,
       bonds = bondsProto(proofOfStake),
       preStateHash = startHash,
       postStateHash = stateHash
@@ -155,87 +144,4 @@ object Genesis {
         Bond(validator, stake)
     }
   }
-
-  private def readBonds[F[_]: Sync: RaiseIOError](
-      bondsPath: Path
-  ): F[Map[PublicKey, Long]] =
-    SourceIO
-      .open[F](bondsPath)
-      .use(_.getLines)
-      .map { lines =>
-        Try {
-          lines
-            .map(line => {
-              val Array(pk, stake) = line.trim.split(" ")
-              PublicKey(Base16.unsafeDecode(pk)) -> stake.toLong
-            })
-            .toMap
-        }
-      }
-      .flatMap {
-        case Success(bonds) => bonds.pure[F]
-        case Failure(_) =>
-          Sync[F].raiseError(new Exception(s"Bonds file $bondsPath cannot be parsed"))
-      }
-
-  def getBonds[F[_]: Sync: Log: RaiseIOError](
-      maybeBondsPath: Option[String],
-      defaultBondsPath: Path,
-      numValidators: Int,
-      genesisPath: Path
-  ): F[Map[PublicKey, Long]] =
-    maybeBondsPath match {
-      case Some(bondsPathStr) =>
-        val bondsPath = Paths.get(bondsPathStr)
-        exists(bondsPath).ifM(
-          readBonds(bondsPath),
-          Sync[F].raiseError(new Exception(s"Specified bonds file $bondsPath does not exist"))
-        )
-      case None =>
-        exists(defaultBondsPath).ifM(
-          Log[F].info(s"Using default file $defaultBondsPath") >> readBonds(defaultBondsPath),
-          Log[F].warn(
-            s"Bonds file was not specified and default bonds file does not exist. Falling back on generating random validators."
-          ) >> newValidators[F](numValidators, genesisPath)
-        )
-    }
-
-  private def newValidators[F[_]: Monad: Sync: Log](
-      numValidators: Int,
-      genesisPath: Path
-  ): F[Map[PublicKey, Long]] = {
-    val keys         = Vector.fill(numValidators)(Secp256k1.newKeyPair)
-    val (_, pubKeys) = keys.unzip
-    val bonds        = pubKeys.zipWithIndex.toMap.mapValues(_.toLong + 1L)
-    val genBondsFile = genesisPath.resolve(s"bonds.txt").toFile
-
-    val skFiles =
-      Sync[F].delay(genesisPath.toFile.mkdir()) >>
-        Sync[F].delay {
-          keys.foreach { //create files showing the secret key for each public key
-            case (sec, pub) =>
-              val sk      = Base16.encode(sec.bytes)
-              val pk      = Base16.encode(pub.bytes)
-              val skFile  = genesisPath.resolve(s"$pk.sk").toFile
-              val printer = new PrintWriter(skFile)
-              printer.println(sk)
-              printer.close()
-          }
-        }
-
-    //create bonds file for editing/future use
-    for {
-      _       <- skFiles
-      printer <- Sync[F].delay { new PrintWriter(genBondsFile) }
-      _ <- bonds.toList.traverse_ {
-            case (pub, stake) =>
-              val pk = Base16.encode(pub.bytes)
-              Log[F].info(s"Created validator $pk with bond $stake") >> Sync[F].delay(
-                printer.println(s"$pk $stake")
-              )
-          }
-      _ <- Sync[F].delay { printer.close() }
-    } yield bonds
-  }
-
 }
